requirements.txt
Django==4.2.2
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
social-auth-app-django==5.2.0
social-auth-core==4.4.2
requests==2.31.0
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
python-telegram-bot==20.4
pytest==7.4.0
